Correctly reuse request body for retries
This package retries requests that fail. That's good!

But it also reused the exact same io.Reader for the request body on the retry.
That's bad! The io.Reader has already been consumed at that point, so it's at EOF.

Fortunately, it turns out that all request bodies were not just generic
io.Readers, but bytes.Buffers. And it's easy to reset and re-read an io.Buffer.
So we do that, and it works!